placed test and api routes in separate folder
