+++ public/pages/template.html
@@ -47,7 +47,7 @@
 		<div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
 		  <ul class="nav navbar-nav">
 		  
-			<li><a href="index.html">Front Page<span class="sr-only">(current)</span></a></li>
+			<li><a href="/pages/index.html">Front Page<span class="sr-only">(current)</span></a></li>
 			
 			<li class="active"><a href="/pages/template.html">Template</a></li>
 			
@@ -168,7 +168,7 @@
 		  </div>
 		  <div class="panel-body">
 		  
-			<form class="navbar-form navbar-left" method="POST" role="form" id="frm1" action="\api\example">
+			<form class="navbar-form navbar-left" method="POST" role="form" id="frm1" action="">
 				
 				<h4>Data in Server <span class="label label-default">New</span></h4>
 				<button type="button" onclick="getDataFromServer()" class="fleft btn btn-secondary">Refresh</button>
